// src/interfaces/notes.interface.ts
export interface ResponseGetNotes {
    msg: string
    notes: Note[]
  }
  
  export interface Note {
    _id: string
    title: string
    content: string
    createdBy: string
    createdAt: string
    updatedAt: string
    __v: number
  }


  export interface DeleteResponse {
    msg: string
  }